backendauramatchnewversion/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
Pillow==10.3.0
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
google-generativeai==0.5.4
dlib==19.24.4
scikit-learn==1.4.2
colour-science==0.4.4
redis==5.0.8
chromadb==0.5.23
sentence-transformers==3.3.1
